Restructure file structure.
